backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
PyMuPDF==1.26.4
PyJWT==2.10.1
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.13
email-validator==2.3.0
